refactor: Centralize network constants and response handling
This commit refactors the network module by centralizing network-related constants and improving the response handling structure.

Key changes:
- **Constants**:
    - Introduced `NetworkConstant` object to hold `BASE_HOST` and `BASE_URL`.
    - Updated `HttpClient` and `provideKtorfit` to use these constants, removing hardcoded URLs.
- **Response Handling**:
    - Renamed `NetworkResult` to `NetworkResultWrapper` and moved it to a dedicated `model` package.
    - Moved the `ResponseHandler` logic into a top-level `handleResponse` function within a new `utils` package.
    - Simplified error handling in `safeBodyAsText` by removing the redundant `CancellationException` check.
