+++ core/network/src/commonMain/kotlin/my/id/tasius/network/Ktorfit.kt
@@ -3,11 +3,9 @@ package my.id.tasius.network
 import de.jensklingenberg.ktorfit.Ktorfit
 import io.ktor.client.HttpClient
 
-private const val DEFAULT_BASE_URL = "https://api.sleep.local/"
-
 fun provideKtorfit(
     httpClient: HttpClient,
-    baseUrl: String = DEFAULT_BASE_URL,
+    baseUrl: String = NetworkConstant.BASE_URL,
 ): Ktorfit =
     Ktorfit.Builder()
         .httpClient(httpClient)
