+++ core/network/src/commonMain/kotlin/my/id/tasius/network/NetworkConstant.kt
@@ -1,0 +1,6 @@
+package my.id.tasius.network
+
+object NetworkConstant {
+    const val BASE_HOST = "localhost:8080"
+    const val BASE_URL = "http://$BASE_HOST"
+}
